Scope Dto classes under ManagerRefresh::
Scope Dto classes under ManagerRefresh::
